Add test for mapping attendance monitoring to inclusion zone
